requirements.txt
Meh==1.1
watchdog
python-telegram-bot==5.2.0
